feat: possibility to force coupled camera
Some mods implements some features that only work well
with a coupled camera, for example, Parcool offers way
to hang on cliffs and turn around. The controls ends
up a bit confused with a decoupled camera. Another example
is lock-on that works pretty well with a coupled camera.
That way, it'd be nice if ShoulderSurfing offered a way
for another mods to force a coupled camera when they need it to.
This, of course, without messing up with the player configuration.
In this commit I'm suggesting a way to do so